+++ list.php
@@ -1,16 +1,25 @@
 <?php
 	include_once("includes/conf.php");
 	include_once("includes/meta.php");
+	include_once("includes/cookies.php");
+	include_once("includes/cache.php");
+	include_once("includes/roles.php");
 	global $SHORT_BASE;
+
+	$session_token = cookies_has_session();
+	$is_admin = roles_is_admin(cache_get_cached_user($session_token));
+
 ?>
 <!DOCTYPE HTML>
 <html lang="en">
 	<head>
 		<title>Redirector URL list</title>
 		<link rel="stylesheet" href="/css/backwards.css">
-		<link rel="icon" href="data:;base64,iVBORw0KGgo=">
+		<link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
+		<link rel="icon" type="image/png" href="/favicon.png" sizes="32x32">
 		<link rel="stylesheet" href="/resources/bootstrap/css/bootstrap.min.css">
 		<link rel="stylesheet" type="text/css" href="//cdn.datatables.net/1.10.11/css/jquery.dataTables.css">
+		<link rel="stylesheet" href="http://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.6.3/css/font-awesome.min.css">
 	</head>
 	<body>
 		<?php include("resources/top_menu.php"); ?>
@@ -25,11 +34,17 @@
 						<th>Long URL</th>
 						<th>Date added</th>
 						<th>Views</th>
+						<?php
+							if($is_admin){
+								echo "<th>Owner</th>";
+								echo "<th>Admin actions</th>";
+							}
+						?>
 					</tr>
 				</thead>
 				<tbody>
 					<?php
-						$sql = "SELECT * FROM translation";
+						$sql = "SELECT short_url, long_url, added, views, u.name AS owner FROM translation t LEFT JOIN users u ON t.owner = u.id";
 						$conn = conf_get_connection();
 						$result = $conn->query($sql);
 						$conn = null;
@@ -43,6 +58,13 @@
 							echo "<td><a href='$long_url_item' target='_blank'>".$item['long_url']."</a></td>";
 							echo "<td>".$item['added']."</td>";
 							echo "<td>".$item['views']."</td>";
+							if($is_admin){
+								echo "<td>".(is_null($item['owner']) ? "-" : $item['owner'])."</td>";
+								echo "<td style='text-align: center'>";
+								echo "<button class='btn btn-primary' data-toggle='modal' data-target='#edit_modal' onclick=\"setEditFields('".$item['long_url']."','".$item['short_url']."')\"><span class='fa fa-pencil'></span></button>&nbsp;&nbsp;&nbsp;";
+								echo "<button class='btn btn-danger' data-toggle='modal' data-target='#confirm_modal' onclick=\"setConfirmFields('".$item['short_url']."')\"><span class='fa fa-trash-o'></span></button>";
+								echo "</td>";
+							}
 							echo "</tr>";
 						}
 					?>
@@ -63,7 +85,22 @@
 					});
 				});
 			</script>
+			<script type="text/javascript">
+				function setEditFields(long_url, short_url) {
+					$("#ed_long_url").val(long_url);
+					$("#ed_short_url").val(short_url);
+					$("#ed_short_url_disabled").val(short_url);
+				}
+				function setConfirmFields(short_url) {
+					$("#conf_short_url").val(short_url);
+					$("#conf_show_url").html(short_url);
+				}
+			</script>
 		</div>
 		<?php include("./resources/footer.php"); ?>
+		<?php
+			include("./resources/edit_modal.html");
+			include("./resources/confirm_modal.html");
+		?>
 	</body>
 </html>